+++ frontend/components/features/article/detail/ArticleContent.vue
