+++ frontend/components/features/article/detail/ArticleHero.vue
